리팩토링 dto에서 또 1:다 dto 구조라 transform -> map으로 해결, fetchjoin미적용 되어 나오는 n+1 error 수정

+++ problem-test/commons/project-dudgns-data-entity/src/main/kotlin/dudgns/com/backend/commons/data/entity/problemTest/CompanyEntity.kt
@@ -10,7 +10,7 @@ import jakarta.persistence.Table
 class CompanyEntity {
     @Id
     @Column(name = "company_code")
-    var companyCode: Int? = null
+    var companyCode: Long? = null
 
     @Column(name = "company_name")
     var companyName: String? = null
